+++ admin/speakers.php
@@ -36,6 +36,11 @@ if (isset($_SESSION['ADM_Username'])){
 	}
 
     if ((isset($_GET['borrar'])) && ($_GET['id'] != "")) {
+	  if (!samap_puede_escribir() || !samap_csrf_validar()) {
+	    echo"<script>alert('No se pudo eliminar el speaker. Volvé a intentarlo.'); window.location.href=\"".$URL."admin/speakers/\"</script>";
+	    exit;
+	  }
+
 	  $deleteSQL = sprintf("DELETE FROM tbl_speaker WHERE id=%s",
 	                       GetSQLValueString($_GET['id'], "int"));
 
@@ -121,7 +126,7 @@ if (isset($_SESSION['ADM_Username'])){
 																					<td><?php echo $row_speakers['nacionalidad'];?></td>
 																					
 																					<td width="20px"><div align="center"><a href="<?php echo $URL?>admin/editarspeaker/cod/<?php echo $row_speakers['id']; ?>/"><img width="20px" src="<?php echo $URL?>admin/app/img/editar.png"alt=""/></a></div></td>
-																					<td width="20px"><div align="center"><a href="<?php echo $URL?>admin/speakers.php?id=<?php echo $row_speakers['id']; ?>&borrar=si"><img width="20px" src="<?php echo $URL?>admin/app/img/borrar.png"alt=""/></a></div></td>
+																					<td width="20px"><div align="center"><a href="<?php echo $URL?>admin/speakers.php?id=<?php echo $row_speakers['id']; ?>&borrar=si&csrf_token=<?php echo urlencode(samap_csrf_valor()); ?>" onclick="return confirm('¿Querés eliminar este speaker? No se puede deshacer.');"><img width="20px" src="<?php echo $URL?>admin/app/img/borrar.png"alt=""/></a></div></td>
 																					
 																				</tr>
                                                                               <?php
